requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.5
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.1.0
django-heroku==0.3.1
django-markdown==0.8.4
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.1
factory-boy==2.11.1
Faker==1.0.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
more-itertools==6.0.0
openapi-codec==1.3.2
pathlib2==2.3.3
Pillow==5.4.1
pluggy==0.8.1
psycopg2==2.7.7
py==1.7.0
pycodestyle==2.5.0
PyGithub==1.43.5
pytest==4.3.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
